+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## unreleased - [v0.0.3]
+* Change module to component module
 
 ## 20118-05-02 - [v0.0.2]
 * Fix composer.json file